requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
openai
pymongo
nest_asyncio
pytz
apscheduler
gunicorn 